Refactor code and create methods to expose functionality that can be reused to implement drag and drop in Windows
